PIM-7629: Rename `CategoryOfAncestorsFilter` to `AggregateFilter` in order to aggregate
