+++ src/Pim/Bundle/CatalogBundle/Resources/config/query_builders.yml
@@ -22,7 +22,6 @@ parameters:
     pim_catalog.query.elasticsearch.filter.id.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\IdFilter
     pim_catalog.query.elasticsearch.filter.family.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\FamilyFilter
     pim_catalog.query.elasticsearch.filter.category.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\CategoryFilter
-    pim_catalog.query.elasticsearch.filter.category_of_ancestors.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\CategoryOfAncestorsFilter
     pim_catalog.query.elasticsearch.filter.group.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\GroupFilter
     pim_catalog.query.elasticsearch.filter.status.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\StatusFilter
     pim_catalog.query.elasticsearch.filter.datetime.class: Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\DateTimeFilter
@@ -285,11 +284,11 @@ services:
             - { name: 'pim_catalog.elasticsearch.query.product_and_product_model_filter', priority: 30 }
 
     pim_catalog.query.elasticsearch.filter.category_of_ancestors:
-        class: '%pim_catalog.query.elasticsearch.filter.category_of_ancestors.class%'
+        class: 'Pim\Bundle\CatalogBundle\Elasticsearch\Filter\Field\AggregateFilter'
         arguments:
         - '@pim_catalog.repository.category'
-        - ['categories_of_ancestors']
-        - ['IN CHILDREN']
+        - ['aggregate']
+        - ['AGGREGATE']
         tags:
         - { name: 'pim_catalog.elasticsearch.query.product_and_product_model_filter', priority: 30 }
